Add new useCallbackState hook, improve bounds calc
The useCallbackState hook allows a callback to use new state without
having to create a new function.
Bounds will now no longer be re-calculated unless the node/edge that
has changed has moved outside of bounds. This may cause the bounds to
become larger than the graph but that's not an issue.